rename code.js to constantUtil.js
